refactor: split migration tests into dedicated file
